backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
bcrypt==4.0.1